+++ packages/fiber/CHANGELOG.md
@@ -1,3 +1,35 @@
 # @react-three/fiber
 
+## 6.2.1
+
+### Patch Changes
+
+- Fix stopPropagation logic
+
+## 6.2.0
+
+### Minor Changes
+
+- Allow object3d instances to be attached
+
+## 6.1.5
+
+### Patch Changes
+
+- fix(rttr): if children is undefined return an array to map with
+
+## 6.1.4
+
+### Patch Changes
+
+- 6faa090: Add shape to types, exclude event functions from event data
+
+## 6.1.3
+
+### Patch Changes
+
+- 71e72c0: Fix constructor args with attached children (#1348)
+- 015fc03: Only set up pointer/wheel events as passive
+- a160e08: Fix event setPointerCapture and stopPropagation.
+
 ## 6.1.2