+++ packages/test-renderer/CHANGELOG.md
@@ -1,3 +1,46 @@
 # @react-three/test-renderer
 
+## 7.0.1
+
+### Patch Changes
+
+- Fix stopPropagation logic
+- Updated dependencies [undefined]
+  - @react-three/fiber@6.2.1
+
+## 7.0.0
+
+### Minor Changes
+
+- Allow object3d instances to be attached
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @react-three/fiber@6.2.0
+
+## 6.1.5
+
+### Patch Changes
+
+- fix(rttr): if children is undefined return an array to map with
+- Updated dependencies [undefined]
+  - @react-three/fiber@6.1.5
+
+## 6.1.3
+
+### Patch Changes
+
+- 6faa090: Add shape to types, exclude event functions from event data
+- Updated dependencies [6faa090]
+  - @react-three/fiber@6.1.4
+
+## 6.1.2
+
+### Patch Changes
+
+- 71e72c0: Fix constructor args with attached children (#1348)
+- Updated dependencies [71e72c0]
+  - @react-three/fiber@6.1.3
+
 ## 6.1.1